YEAST! I don't know why I am writing this but I am...
I guess I wanted something special for the first post. Either way, I
committed it earlier but it may not have been pushed, so here you go.
And I edited the first paragraph a bit so it sounds more uniform. I will
also make the excerpt a lot shorter in the future, and of course, finish
the post itself.