Add files via upload
Wrote this a while ago in Gitpod but had some problems with committing it (weird Gitpod related error), so I couldn't get it to save here until today. Also I've been really busy so the info in the post is completely wrong. I'm  currently writing a post on what's really going on, not incorrect info. I'm also doing a million other things that will be explained in that future post.